backend/requirements.txt
fastapi
strawberry-graphql[fastapi]
psycopg[binary]
sqlalchemy
uvicorn
redis
passlib[bcrypt]
python-jose[cryptography]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-redis
prometheus_client
structlog
google-auth-oauthlib
httpx
fastapi-limiter
aiocache
